+++ tests/test_state_props/test_is_ppt.py
@@ -2,7 +2,7 @@
 import numpy as np
 
 from toqito.state_props import is_ppt
-from toqito.states import bell
+from toqito.states import bell, horodecki
 
 
 def test_is_ppt():
@@ -33,6 +33,11 @@ def test_entangled_state():
     """Entangled state of dimension 2 will violate PPT criterion."""
     rho = bell(2) * bell(2).conj().T
     np.testing.assert_equal(is_ppt(rho), False)
+
+
+def test_is_horodecki_ppt():
+    """Horodecki state is an example of an entangled PPT state."""
+    np.testing.assert_equal(is_ppt(horodecki(0.5, [3, 3])), True)
 
 
 if __name__ == "__main__":