+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 
 - Fix: Bug in `partial_transpose` for non-square matrices. Adding test cases to
   cover this previously failing case.
+ 
+- Fix: Bug in `purity`. Squaring matrix was using `**2` instead of 
+  `np.linalg.matrix_power` 
 
 - Feature: Added in `symmetric_extension_hierarchy`. The function is a
   hierarchy of semidefinite programs that eventually converge to the separable
@@ -25,6 +28,10 @@
 - Feature: Added in ability to perform both minimum-error and unambiguous state
   discrimination in `state_distinguishability.py`. Adding additional tests to
   cover this extra feature.
+  
+- Feature: Added `majorizes`; a function to determine whether a vector or matrix
+  majorizes another. This is used as a criterion to determine whether a quantum
+  state can be converted to another via LOCC.
 
 - Enhancement: Adding further tests for `symmetric_projection`.